PLANET-7745: Adjust taxonomy breadcrumbs
- Add the post type as attribute of the new block
