+++ assets/src/editorIndex.js
@@ -18,7 +18,7 @@ import {addButtonLinkPasteWarning} from './block-editor/addButtonLinkPasteWarnin
 import {addBlockFilters} from './block-editor/BlockFilters';
 import {replaceTaxonomyTermSelectors} from './block-editor/replaceTaxonomyTermSelectors';
 import {setupImageBlockExtension} from './block-editor/setupImageBlockExtension';
-import {setupPostsListBlockExtension} from './block-editor/setupPostsListBlockExtension';
+import {setupTaxonomyBreadcrumbBlock} from './block-editor/setupTaxonomyBreadcrumbBlock';
 
 wp.domReady(() => {
   // Blocks
@@ -55,4 +55,4 @@ addBlockFilters();
 replaceTaxonomyTermSelectors();
 setupImageBlockExtension();
 setupBlockEditorValidation();
-setupPostsListBlockExtension();
+setupTaxonomyBreadcrumbBlock();
